more trimming/cleanup. simple test
